20230421_PoisonAttack/requirement.txt
torch==1.4.0 torchvision==0.5.0
pysyft==0.2.4
tqdm
scikit-learn
pandas
PyYaml